refactor(did): split DID resolvers into jwk/web files and add CompositeResolver cache
